Merge pull request #92 from insolite-dev/develop

+++ lib/commands/remove.go
@@ -8,6 +8,7 @@ package commands
 
 import (
 	"fmt"
+	"os"
 
 	"github.com/AlecAivazis/survey/v2"
 	"github.com/insolite-dev/notya/assets"
@@ -59,7 +60,7 @@ func runRemoveCommand(cmd *cobra.Command, args []string) {
 	loading.Start()
 
 	// Generate array of all node names.
-	_, nodeNames, err := service.GetAll("", models.NotyaIgnoreFiles)
+	_, nodeNames, err := service.GetAll("", "", models.NotyaIgnoreFiles)
 
 	loading.Stop()
 	if err != nil {
@@ -79,6 +80,11 @@ func runRemoveCommand(cmd *cobra.Command, args []string) {
 
 // removeAndFinish removes given node and alerts success message if everything is OK.
 func removeAndFinish(node models.Node) {
+	if len(node.Title) == 0 {
+		os.Exit(-1)
+		return
+	}
+
 	loading.Start()
 
 	err := service.Remove(node)
